FIX: prevent session clobbering
This did not seem to happen when long-polling, but simple polling was
sending a session cookie back to the browser. MessageBus polling should
not have side effects other than handling messages. In particular, it
should not be sending session cookies, as this can clobber session
state set by other means.

We fix the issue by setting the :skip Rack session option to true, which
prevents the Rack session middleware from committing the session cookie
to headers.